+++ gui/src/test/java/io/xj/gui/services/FabricationServiceImplTest.java
@@ -41,7 +41,7 @@ class FabricationServiceImplTest {
   private final String defaultMacroMode = MacroMode.AUTO.toString();
   String defaultInputMode = InputMode.PRODUCTION.toString();
   String defaultOutputFileMode = OutputFileMode.CONTINUOUS.toString();
-  float defaultOutputFrameRate = 48000;
+  double defaultOutputFrameRate = 48000;
   String defaultOutputMode = OutputMode.PLAYBACK.toString();
   int defaultOutputSeconds = 300;
 
@@ -107,18 +107,18 @@ class FabricationServiceImplTest {
   @Test
   void formatPositionBarBeats() throws ManagerFatalException {
     var segment4 = prepareSegmentManager(4);
-    assertEquals("1.1", subject.formatPositionBarBeats(segment4, 0.0f));
-    assertEquals("2.1", subject.formatPositionBarBeats(segment4, 4.0f));
-    assertEquals("2.2", subject.formatPositionBarBeats(segment4, 5.0f));
-    assertEquals("2.3", subject.formatPositionBarBeats(segment4, 6.0f));
-    assertEquals("2.3.5", subject.formatPositionBarBeats(segment4, 6.5f));
-    assertEquals("2.3.523", subject.formatPositionBarBeats(segment4, 6.523f));
-    assertEquals("3.1", subject.formatPositionBarBeats(segment4, 8.0f));
-    assertEquals("4.1", subject.formatPositionBarBeats(segment4, 12.0f));
+    assertEquals("1.1", subject.formatPositionBarBeats(segment4, 0.0));
+    assertEquals("2.1", subject.formatPositionBarBeats(segment4, 4.0));
+    assertEquals("2.2", subject.formatPositionBarBeats(segment4, 5.0));
+    assertEquals("2.3", subject.formatPositionBarBeats(segment4, 6.0));
+    assertEquals("2.3.5", subject.formatPositionBarBeats(segment4, 6.5));
+    assertEquals("2.3.523", subject.formatPositionBarBeats(segment4, 6.523));
+    assertEquals("3.1", subject.formatPositionBarBeats(segment4, 8.0));
+    assertEquals("4.1", subject.formatPositionBarBeats(segment4, 12.0));
 
     var segment3 = prepareSegmentManager(3);
-    assertEquals("5.1", subject.formatPositionBarBeats(segment3, 12.0f));
-    assertEquals("5.3.5", subject.formatPositionBarBeats(segment3, 14.5f));
+    assertEquals("5.1", subject.formatPositionBarBeats(segment3, 12.0));
+    assertEquals("5.3.5", subject.formatPositionBarBeats(segment3, 14.5));
   }
 
   private Segment prepareSegmentManager(int barBeats) throws ManagerFatalException {
